diplomovka/requirements.txt
pandas
numpy
scikit-learn
nltk
stanza
xgboost
python-docx
joblib
